move query tests into their own package
