+++ query/tests/command_names_test.go
@@ -13,12 +13,13 @@
 // limitations under the License.
 
 // Integration test for the query execution.
-package query
+package tests
 
 import (
 	"testing"
 
 	"github.com/square/metrics/api"
+	"github.com/square/metrics/query/command"
 	"github.com/square/metrics/query/parser"
 	"github.com/square/metrics/testing_support/mocks"
 )
@@ -161,7 +162,7 @@ func TestQueryNaming(t *testing.T) {
 			t.Errorf("Expected select command but got %s", testCommand.Name())
 			continue
 		}
-		rawResult, err := testCommand.Execute(ExecutionContext{
+		rawResult, err := testCommand.Execute(command.ExecutionContext{
 			TimeseriesStorageAPI: fakeBackend,
 			MetricMetadataAPI:    fakeAPI,
 			FetchLimit:           1000,
@@ -171,7 +172,7 @@ func TestQueryNaming(t *testing.T) {
 			t.Errorf("Unexpected error while execution: %s", err.Error())
 			continue
 		}
-		seriesListList, ok := rawResult.Body.([]QuerySeriesList)
+		seriesListList, ok := rawResult.Body.([]command.QuerySeriesList)
 		if !ok || len(seriesListList) != 1 {
 			t.Errorf("expected query `%s` to produce []QuerySeriesList; got %+v :: %T", test.query, rawResult.Body, rawResult.Body)
 			continue
